app/requirements.txt
flask==2.0.2
mysql-connector-python
#mysql.connector==2.2.9